+++ docs/CHANGELOG.md
@@ -1,4 +1,16 @@
-## [0.3.1]
+## [0.4.0] - 2026-04-30
+
+### Changes
+
+- Major code overhaul
+- Better model, view, controller structure
+- Moved logic out of view.ts into engine.ts
+- Moved interfaces to types.ts
+- Moved YAML config into plugin settings
+- sidebar.ts is now dashboard-leaf.ts and opens in main window, not sidebar
+- Responsiveness fixes and style changes
+
+## [0.3.1] - 2026-04-30
 
 ### Changes